ui: widen the cell index to int32_t - windows past 32767 cells draw
the screen and window cell indices were int16_t, so any window larger
than 32767 cells wrapped negative and the draw walked off the front of
the buffer.  uint16_t only moves the wall to 65535; the products are
what overflow, so the indices widen to int32_t and every multiplication
is cast before it happens.  the dimensions stay int16_t - only their
products grow.

uC_window_draw.c's index was worse than the others: a BYTE count, so it
overflowed at a width of 4096 rather than 32767.

verified against the installed library: 132 doc/size pairs and the axes
suite pass, and every size over 32767 cells now draws.

+++ src/ui/uC_border.c
@@ -91,7 +91,7 @@ static void draw_char(uC_window_t *win, int16_t cx, int16_t cy,
 {
     uC_screen_t *scr = win->screen;
 
-    int16_t index = (cy * scr->width) + cx;
+    int32_t index = ((int32_t)cy * (int32_t)scr->width) + cx;
 
     cell_t *p1 = &scr->buffer1[index];
     cell_t *p2 = &scr->buffer2[index];
